fix(db): add unique constraint on iris to avoid duplicate

+++ app-near/scripts/postdeploy.sh
@@ -2,15 +2,17 @@
 
 npm run prisma migrate deploy
 
-curl -L -O https://www.insee.fr/fr/statistiques/fichier/8268806/base-ic-evol-struct-pop-2021_csv.zip
+if [[ $IS_REVIEW_APP = "true" ]]; then
+    curl -L -O https://www.insee.fr/fr/statistiques/fichier/8268806/base-ic-evol-struct-pop-2021_csv.zip
 
-unzip base-ic-evol-struct-pop-2021_csv.zip
+    unzip base-ic-evol-struct-pop-2021_csv.zip
 
-mv base-ic-evol-struct-pop-2021.CSV ./scripts/base-ic-evol-struct-pop-2021.csv
+    mv base-ic-evol-struct-pop-2021.CSV ./scripts/base-ic-evol-struct-pop-2021.csv
 
-rm base-ic-evol-struct-pop-2021_csv.zip
+    rm base-ic-evol-struct-pop-2021_csv.zip
 
-npm run load:insee -- "./scripts/base-ic-evol-struct-pop-2021.csv" ${SCALINGO_POSTGRESQL_URL}
+    npm run load:insee -- "./scripts/base-ic-evol-struct-pop-2021.csv" ${SCALINGO_POSTGRESQL_URL}
+fi
 
 npm run load:surveys
